Refactor: split large UI files into focused modules
Break down the 5 largest source files (968–1315 lines each) into
smaller, single-responsibility modules to improve navigability and
maintainability. Pure file reorganization — no logic changes.

- SettingsSheet.kt (968→114): extract SettingsComponents, SoundSection,
  DisplaySection, InstrumentSettingsSections, AboutSection
- ScalePracticeView.kt (942→180): extract PlayAlong, Quiz,
  EarTraining, Stats composables
- ChordLibraryTab.kt (978→311): extract ChordSearchBar, ChordSelectors,
  InversionCompareView, VoicingGrid
- TunerViewModel.kt (902→843): extract TunerUiState data class + enums
- SongbookView.swift (1315→1220): extract PerformanceModeView
